Update appState store
