+++ src/store/scores/actions.ts
@@ -3,7 +3,7 @@ import { Action, Dispatch } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 
 import { API_URL } from '../../config/constants';
-import { appDoneLoading, appLoading, setMessage } from '../appState/actions-creators';
+import { appDoneLoading, appLoading, setMessage } from '../appState/actions';
 import { StoreState } from '../types';
 import {
   FixtureWithScores,
